Added alpha letter to listing
Before the list of names which are alphabetical by last name as determined by the url, the first letter of the last name is displayed before each section.  So before the last names beginning with "a", a large "A" is displayed in the row before the first listing.  Continues for each letter where there exists last names.